created folder for new Footer component and styles for desktop design
