+++ src/pages/IA.jsx
@@ -4,7 +4,7 @@ import WhySection from '../components/IA-salud/WhySection';
 import MiIdentidadVocal from '../components/IA-salud/MiIdentidadVocal';
 import NewsSection from '../components/IA-salud/NewsSection';
 import { IaWrapper } from '../styles/pages/ia.style';
-import Footer from '../components/Footer';
+import Footer from '../components/footer/Footer';
 
 function IA() {
   return (
